% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listHIE.R
\name{listHIE}
\alias{listHIE}
\title{Find data from HIE athletes.}
\usage{
listHIE(folderToScan)
}
\arguments{
\item{folderToScan}{Path for the folder containing the datas.}
}
\value{
Two dataframes listing the files for HIE and not HIE athletes and their oxygen saturation difference (SaO2)
}
\description{
Find data from HIE athletes.
}
